+++ virtio-devices/src/vsock/mod.rs
@@ -285,9 +285,9 @@ mod tests {
         pub fn create_epoll_handler_context(&self) -> EpollHandlerContext {
             const QSIZE: u16 = 2;
 
-            let guest_rxvq = GuestQ::new(GuestAddress(0x0010_0000), &self.mem, QSIZE as u16);
-            let guest_txvq = GuestQ::new(GuestAddress(0x0020_0000), &self.mem, QSIZE as u16);
-            let guest_evvq = GuestQ::new(GuestAddress(0x0030_0000), &self.mem, QSIZE as u16);
+            let guest_rxvq = GuestQ::new(GuestAddress(0x0010_0000), &self.mem, QSIZE);
+            let guest_txvq = GuestQ::new(GuestAddress(0x0020_0000), &self.mem, QSIZE);
+            let guest_evvq = GuestQ::new(GuestAddress(0x0030_0000), &self.mem, QSIZE);
             let rxvq = guest_rxvq.create_queue();
             let txvq = guest_txvq.create_queue();
             let evvq = guest_evvq.create_queue();
